ros2_ws/requirements.txt
action-msgs==1.2.2
action-tutorials-interfaces==0.20.8
action-tutorials-py==0.20.8
actionlib-msgs==4.9.1
ament-cmake-test==1.3.14
ament-copyright==0.12.15
ament-cppcheck==0.12.15
ament-cpplint==0.12.15
ament-flake8==0.12.15
ament-index-python==1.4.1
ament-lint==0.12.15
ament-lint-cmake==0.12.15
ament-package==0.14.2
ament-pep257==0.12.15
ament-uncrustify==0.12.15
ament-xmllint==0.12.15
angles==1.15.0
appdirs==1.4.4
apturl==0.5.2
argcomplete==1.8.1
attrs==21.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
bond==4.1.4
Brlapi==0.8.3
Brotli==1.0.9
builtin-interfaces==1.2.2
catkin-pkg-modules==1.1.0
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.20.1
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.4
command-not-found==0.3
composition-interfaces==1.2.2
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cupshelpers==1.0
cv-bridge==3.2.1
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
demo-nodes-py==0.20.8
diagnostic-msgs==4.9.1
diagnostic-updater==4.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.17.1
domain-coordinator==0.10.0
duplicity==0.8.21
dwb-msgs==1.1.20
empy==3.3.4
example-interfaces==0.9.3
examples-rclpy-executors==0.15.5
examples-rclpy-minimal-action-client==0.15.5
examples-rclpy-minimal-action-server==0.15.5
examples-rclpy-minimal-client==0.15.5
examples-rclpy-minimal-publisher==0.15.5
examples-rclpy-minimal-service==0.15.5
examples-rclpy-minimal-subscriber==0.15.5
fasteners==0.14.1
filelock==3.25.2
flake8==4.0.1
fonttools==4.29.1
fs==2.4.12
fsspec==2026.2.0
future==0.18.2
gast==0.5.2
geographic-msgs==1.0.6
geometry-msgs==4.9.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
image-geometry==3.2.1
importlib-metadata==4.6.4
iniconfig==1.1.1
interactive-markers==2.3.2
jeepney==0.7.1
Jinja2==3.1.6
joint-state-publisher==2.4.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
lark==1.1.1
laser-geometry==2.4.1
launch==1.0.14
launch-ros==0.19.13
launch-testing==1.0.14
launch-testing-ros==0.19.13
launch-xml==1.0.14
launch-yaml==1.0.14
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lifecycle-msgs==1.2.2
lockfile==0.12.2
logging-demo==0.20.8
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
map-msgs==2.1.0
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
message-filters==4.3.14
monotonic==1.6
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==1.3.0
nav-2d-msgs==1.1.20
nav-msgs==4.9.1
nav2-common==1.1.20
nav2-msgs==1.1.20
nav2-simple-commander==1.0.0
netifaces==0.11.0
networkx==3.4.2
nose2==0.9.2
notify2==0.3
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
opencv-python==4.13.0.92
osrf-pycommon==2.1.6
packaging==21.3
paramiko==2.9.3
pcl-msgs==1.0.0
pendulum-msgs==0.20.8
pexpect==4.8.0
Pillow==9.0.1
pluggy==0.13.0
ply==3.11
polars==1.39.2
polars-runtime-32==1.39.2
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.10.0
pycairo==1.20.1
pycodestyle==2.8.0
pycups==2.0.1
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pyserial==3.5
pytesseract==0.3.13
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu4.1
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-qt-binding==1.1.3
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.8
rcl-interfaces==1.2.2
rclpy==3.3.21
rcutils==5.1.8
reportlab==3.6.8
requests==2.25.1
resource-retriever==3.1.3
rmw-dds-common==1.6.0
robot-localization==3.5.4
roman==3.3
ros2action==0.18.18
ros2bag==0.15.16
ros2cli==0.18.18
ros2component==0.18.18
ros2doctor==0.18.18
ros2interface==0.18.18
ros2launch==0.19.13
ros2lifecycle==0.18.18
ros2multicast==0.18.18
ros2node==0.18.18
ros2param==0.18.18
ros2pkg==0.18.18
ros2plugin==5.1.3
ros2run==0.18.18
ros2service==0.18.18
ros2topic==0.18.18
rosbag2-interfaces==0.15.16
rosbag2-py==0.15.16
rosdistro-modules==1.0.1
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.8
rosidl-cli==3.1.8
rosidl-cmake==3.1.8
rosidl-generator-c==3.1.8
rosidl-generator-cpp==3.1.8
rosidl-generator-py==0.14.6
rosidl-generator-rs==0.4.11
rosidl-parser==3.1.8
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.4
rosidl-typesupport-fastrtps-cpp==2.2.4
rosidl-typesupport-introspection-c==3.1.8
rosidl-typesupport-introspection-cpp==3.1.8
rospkg-modules==1.6.1
rover-arm==0.0.0
rover-bringup==0.0.0
rover-control1==0.0.0
rover-mapping==0.0.0
rover-mission==0.0.0
rover-monitor==0.0.0
rover-navigation==0.0.0
rover-vision==0.0.0
RPi.GPIO==0.7.1a4
rpyutils==0.2.2
rqt-action==2.0.1
rqt-bag==1.1.6
rqt-bag-plugins==1.1.6
rqt-console==2.0.3
rqt-graph==1.3.2
rqt-gui==1.1.9
rqt-gui-py==1.1.9
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.9
rqt-py-console==1.0.2
rqt-reconfigure==1.1.4
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.1
scipy==1.8.0
SecretStorage==3.3.1
sensor-msgs==4.9.1
sensor-msgs-py==4.9.1
shape-msgs==4.9.1
six==1.16.0
slam-toolbox==2.6.10
smclib==4.1.4
snowballstemmer==2.2.0
soupsieve==2.3.1
sros2==0.10.9
statistics-msgs==1.2.2
std-msgs==4.9.1
std-srvs==4.9.1
stereo-msgs==4.9.1
sympy==1.14.0
systemd-python==234
teleop-twist-keyboard==2.4.1
tf-transformations==1.1.1
tf2-geometry-msgs==0.25.19
tf2-kdl==0.25.19
tf2-msgs==0.25.19
tf2-py==0.25.19
tf2-ros-py==0.25.19
tf2-sensor-msgs==0.25.19
tf2-tools==0.25.19
toml==0.10.2
topic-monitor==0.20.8
trajectory-msgs==4.9.1
transforms3d==0.3.1
turtlesim==1.4.3
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
ultralytics-thop==2.0.18
unattended-upgrades==0.1
unicodedata2==14.0.0
unique-identifier-msgs==2.2.1
urllib3==1.26.5
visualization-msgs==4.9.1
wadllib==1.3.6
webencodings==0.5.1
xacro==2.1.1
xdg==5
xkit==0.0.0
zipp==1.0.0
